se agregaron los hooks useffect

+++ src/components/Ventas.jsx
@@ -1,57 +1,59 @@
 import React, { useEffect } from 'react'
-//import { useEffect } from 'react';
 import Table from 'react-bootstrap/Table';
-//import { Footer } from '../../layout/Footer';
 import Button from 'react-bootstrap/Button';
-//import { useNavigate } from 'react-router';
 import  { useState } from 'react'
 import { Container } from 'react-bootstrap';
 import { NavbarMain } from './NavbarMain';
-import { ModalEditarProducto } from './ModalEditarProducto';
-
 import sigecoApi from '../api/sigecoAPI';
+import Swal from 'sweetalert2';
 
 export const Ventas = () => {
 
    const [ventasMostradas, setventasMostradas] = useState([]);
-   const [modalAbierto, setModalAbierto] = useState(false);
-   const [modalAbiertoAP, setModalAbiertoAP] = useState(false);
-//    const  [formDataEdit,setFormDataEdit ]= useState({
-//     _id:"",
-//     detalle:[],
-//     total:"",
-//     fecha: "",
+   const [searchTerm, setSearchTerm] = useState('');
+   const [cantidadVentas, setcantidadVentas] = useState(0);
+   const [sumaVentas, setsumaVentas] = useState(0);
+
+   useEffect(() => {
  
-// })
-
-
-const [searchTerm, setSearchTerm] = useState('');
-
-const abrirModal = () => setModalAbierto(true);
-const cerrarModal=() => setModalAbierto(false)
-
-const abrirModalAP = () => setModalAbiertoAP(true);
-const cerrarModalAP=() => setModalAbiertoAP(false)
-
-
-    const ventasDB= async()=>{ try {
+    ventasDB();
+  }, []);
+  useEffect(() => {
+    // Esto se ejecutará después de que se haya actualizado ventasMostradas
+    sumaTotal();
+  }, [ventasMostradas]); 
+  
+const ventasDB= async()=>{ try {
       const resp = await sigecoApi.get("/api/ventas")
       setventasMostradas(resp.data.ventas);
-      
+       
     } catch (error) {
       console.log(error);
     }
   }
-
-    //  const editarProductoClick = (producto) => {
-    //   abrirModal();
-    //     setFormDataEdit(producto);
       
-    //   }
-      
-      const eliminarVentaDB = async(id) => {
+const eliminarVentaDB = async(id) => {
         try{
-          const resp=await sigecoApi.delete(`/api/ventas/${id}`);
+          Swal.fire({
+            title: "Está seguro que desea eliminar esta venta?",
+            
+            showCancelButton: true,
+            confirmButtonText: "Eliminar",
+            
+          }).then (async(result) => {
+            /* Read more about isConfirmed, isDenied below */
+            if (result.isConfirmed) {
+              const resp=await sigecoApi.delete(`/api/ventas/${id}`);
+              Swal.fire("venta elminada", "", "success");
+              ventasDB();
+            } else if (result.isDenied) {
+              Swal.fire("Operación cancelada", "", "info");
+            }
+          });
+
+
+
+   
        
           
       }
@@ -66,7 +68,7 @@ const cerrarModalAP=() => setModalAbiertoAP(false)
       }
       };
 
-      const handleSearch = (event) => {
+const handleSearch = (event) => {
         if (event.key === 'Enter')
         {
           
@@ -77,6 +79,7 @@ const cerrarModalAP=() => setModalAbiertoAP(false)
               //||   venta.codigo.toString().includes(searchTerm)
           );
           setventasMostradas(ventasFiltradas);
+          sumaTotal();
 
         }
         else {
@@ -85,67 +88,68 @@ const cerrarModalAP=() => setModalAbiertoAP(false)
        
       };
 
+const sumaTotal=()=>{
+    const cantidad=  ventasMostradas.length;
+    const suma = ventasMostradas.reduce((acumulador, venta) => acumulador + venta.total, 0);
+    setcantidadVentas(cantidad);
+    setsumaVentas(suma);
+     };
      
-      return (
+     
+     return (
     <div>
      
 <NavbarMain/>
 <Container className='m-4'>
 <h2 className='text-center'>VENTAS</h2>
-
-
+<h5>Cantidad de ventas:{cantidadVentas}</h5>
+<h5>Recaudacion total:{sumaVentas}</h5>
 <Button className='ms-2' onClick={()=>ventasDB()}>Todas las ventas</Button>
 <input className='ms-2'
         type="text"
-        placeholder="Buscar por nombre o id"
+        placeholder="Buscar por fecha"
         value={searchTerm}
         onChange={handleSearch}
         onKeyDown={handleSearch}/>
 
 <Table striped bordered hover className='m-3'>
-      <thead>
-        <tr>
-          <th>Detalle</th>
-          <th>Total</th>
-          <th>Fecha</th>
-          <th>Eliminar</th>
-          {/* <th>Editar</th> */}
+  <thead>
+    <tr>
+      <th>#</th>
+      <th>Detalle</th>
+      <th>Total</th>
+      <th>Fecha</th>
+      <th>Eliminar</th>
+    </tr>
+  </thead>
+  <tbody>
+    {ventasMostradas.map((venta, index) => {
+      return (
+        <tr key={venta._id}>
+          <td>{index + 1}</td>
+          <td>
+            <h6>Productos:</h6>
+            {venta.detalle.map((item) => (
+              <p>
+                {item.nombre}, <b>Cant:</b> {item.cantidad}
+              </p>
+            ))}
+          </td>
+          <td>${venta.total}</td>
+          <td>{venta.fecha}</td>
+          <td>
+            <Button onClick={() => eliminarVentaDB(venta._id)} variant='danger'>
+              Eliminar
+            </Button>
+          </td>
         </tr>
-      </thead>
-      <tbody>
-{ventasMostradas.map((venta) =>{
-return(
-
-    <tr key={venta._id}>
-    <td>
-    <h6>Productos:</h6>
-        {venta.detalle.map((item) =>{ return ( <p> {item.nombre} , <b>Cant:</b> {item.cantidad} </p>
-           )     })}
-        
-      
-        </td>
-    <td>{venta.total}</td>
-    <td>{venta.fecha}</td>
-   
-    <td><Button onClick={()=> eliminarVentaDB(product._id)} variant='danger'>Eliminar</Button ></td>
-    {/* <td><Button onClick={()=> editarProductoClick(product)} variant='info'>Editar</Button ></td> */}
-  </tr>
-)
-
-})}
-    </tbody>
-    </Table>
+      );
+    })}
+  </tbody>
+</Table>
 
 
 </Container>
-
-
-{/* 
-{modalAbierto && <ModalEditarProducto  datosProducto={formDataEdit}  cerrarModal={cerrarModal}    />}
-
-{modalAbiertoAP && <ModalAgregarProducto    cerrarModal={cerrarModalAP}    />} */}
-
-
 
 
     </div>
